refactor(chat): extract message bubbles
